reverting cloudfront redirect approach for wag-421 (see notes)
It turns out that this fix inadvertently breaks the legacy
documents redirect in website/app/urls.py. This is because the
cloudfront function will intercept all requests that start with
'/resources/', but only adds one redirect (for rule-27.pdf).
There are hundreds of documents from the legacy website that
used this url pattern for documents, and still several sites
that link to those legacy documents. So reverting the files
seems like the better move for now.