renaming migration 71 to 70 to resolve conflicts (see notes)
our migration file names somehow got deviated between the main branch
and the feature branch, such that we had 0070_newsitem_note_newsitem_review_by
in main branch and 0071_newsitem_note_newsitem_review_by in feature branch.

this change attempts to right things by renaming the file (and updating
references to said file) to 0070_newsitem_note_newsitem_review_by in
feature branch so that this migration is ignored and the rest of
the pending migrations can proceed.